+++ test/endpoint.js
@@ -124,6 +124,7 @@ describe('stemmers', function() {
 describe('structures', function() {
   require('./structures/bk-tree.js');
   require('./structures/bloom-filter.js');
+  // require('./structures/symspell.js');
 });
 
 describe('tag', function() {