+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Adding the `parsers/conll` namespace.
 * Adding the `structures/bk-tree` namespace.
 * Adding the `structures/bloom-filter` namespace.
+* Adding the `structures/symspell` namespace.
 * Adding the `sampler` options to `clustering/k-means`.
 * Adding the `stats/descriptive#.quantile` function.
 * Adding the `stats/descriptive#.median` function.